tolk-js v0.6 implementation
tolk-js is conceptually similar to func-js.
But unlike func-js, which is separated into two packages (func-js-bin),
tolk-js contains API and WASM here, in a single place.
Its versioning will be maintained equal to a Tolk Compiler,
since it's an essential part of distribution.

Historically, tolk-js was developed for several months,
in sync with the development of the Tolk compiler.
But for the first public release, that dev history is not needed.
That's why I just do a single giant commit.